finished post form and moved post list to post folder
